+++ app/src/main/AndroidManifest.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     package="com.ericabraham.leapfrog">
 
   <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION"/>
@@ -13,12 +14,15 @@
       android:label="@string/app_name"
       android:roundIcon="@mipmap/ic_launcher_round"
       android:supportsRtl="true"
-      android:theme="@style/AppTheme">
-    <service android:name=".GeofenceTrasitionService"/>
+      android:theme="@style/AppTheme"
+      android:name=".MyLeapFrogApplication"
+      android:appComponentFactory = "androidx"
+      tools:replace="android:appComponentFactory">
+    <service android:name=".Ui.GeofenceTrasitionService"/>
 
-    <receiver android:name=".MainActivity$NotificationMgr"/>
-    <receiver android:name=".MainActivity$SkipTask"/>
-    <receiver android:name=".MainActivity$MarkDone"/>
+    <receiver android:name=".Ui.MainActivity$NotificationMgr"/>
+    <receiver android:name=".Ui.MainActivity$SkipTask"/>
+    <receiver android:name=".Ui.MainActivity$MarkDone"/>
 
     <meta-data
         android:name="com.google.android.geo.API_KEY"
@@ -28,7 +32,7 @@
         android:value="@integer/google_play_services_version"/>
 
     <activity
-        android:name=".Splash"
+        android:name=".Ui.Splash"
         android:label="@string/app_name"
         android:screenOrientation="portrait"
         android:theme="@style/AppTheme.NoActionBar">
@@ -39,19 +43,19 @@
       </intent-filter>
     </activity>
     <activity
-        android:name=".TaskSetting"
+        android:name=".Ui.TaskSetting"
         android:screenOrientation="portrait"/>
     <activity
-        android:name=".ManageTask"
+        android:name=".Ui.ManageTask"
         android:screenOrientation="portrait"/>
     <activity
-        android:name=".MainActivity"
+        android:name=".Ui.MainActivity"
         android:screenOrientation="portrait"/>
     <activity
-        android:name=".About"
+        android:name=".Ui.About"
         android:screenOrientation="portrait"/>
     <activity
-        android:name=".MyMap"
+        android:name=".Ui.MyMap"
         android:screenOrientation="portrait"/>
 
   </application>
