+++ app/src/main/java/com/ericabraham/leapfrog/Interface/RefreshReminderList.java
@@ -1,0 +1,6 @@
+package com.ericabraham.leapfrog.Interface;
+
+public interface RefreshReminderList {
+
+  public void onRefreshReminderList();
+}
