+++ app/src/main/java/com/ericabraham/leapfrog/MyLeapFrogApplication.java
@@ -1,0 +1,12 @@
+package com.ericabraham.leapfrog;
+
+import android.app.Application;
+import android.support.multidex.MultiDex;
+
+public class MyLeapFrogApplication extends Application {
+
+  @Override public void onCreate() {
+    super.onCreate();
+    MultiDex.install(this);
+  }
+}
