+++ app/src/main/java/com/ericabraham/leapfrog/Ui/ManageTask.java
@@ -1,4 +1,4 @@
-package com.ericabraham.leapfrog;
+package com.ericabraham.leapfrog.Ui;
 
 import android.app.DatePickerDialog;
 import android.content.Intent;
@@ -9,29 +9,33 @@ import android.widget.Button;
 import android.widget.DatePicker;
 import android.widget.EditText;
 import android.widget.SeekBar;
+import android.widget.TextView;
 import android.widget.Toast;
 
+import com.ericabraham.leapfrog.Database.locationDatabase;
+import com.ericabraham.leapfrog.R;
+import com.ericabraham.leapfrog.Utils.Utils;
 import java.util.Calendar;
 
-
-public class TaskSetting extends AppCompatActivity {
+public class ManageTask extends AppCompatActivity {
 
     private EditText date;
     private DatePickerDialog datePickerDialog;
-    private static final String TAG = "TaskSetting";
+    private int idno;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_task_setting);
+        setContentView(R.layout.activity_manage_task);
 
         // initiate the date picker and a button
         date = findViewById(R.id.date);
         date.setKeyListener(null);
-        // perform click event on edit text
 
+        // perform click event on edit text
         date.setOnFocusChangeListener(new View.OnFocusChangeListener() {
             @Override public void onFocusChange(View v, boolean hasFocus) {
+                // calender class's instance and get current date , month and year from calender
                 if (hasFocus) {
                     onDatePickerDialog();
                 }
@@ -44,73 +48,72 @@ public class TaskSetting extends AppCompatActivity {
             }
         });
 
+        int id = (int) getIntent().getExtras().get("taskID");
+        //   Toast.makeText(getApplicationContext(),"Ahhh..I am Manage Task, i got: " +i ,Toast.LENGTH_SHORT).show();
+
+
         final locationDatabase db = new locationDatabase(this);
 
-        Button cancel_button = findViewById(R.id.cancel_button);
+        final TextView title = findViewById(R.id.title_edittext);
+        final SeekBar radius = findViewById(R.id.seekBar);
+        final TextView date = findViewById(R.id.date);
+        final TextView todo = findViewById(R.id.todo_edittext);
+        TextView tid = findViewById(R.id.taskid);
 
-        cancel_button.setOnClickListener(new View.OnClickListener() {
+
+        String[] taskData = db.displayTask(id);
+        int raduisData = db.displayRadius(id);
+
+        title.setText(taskData[0]);
+        radius.setProgress(raduisData);
+        date.setText(taskData[1]);
+        todo.setText(taskData[2]);
+        tid.setText(taskData[3]);
+        idno = Integer.valueOf(taskData[3]);
+
+
+        Button del_button = findViewById(R.id.del_button);
+
+        del_button.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-
-                //   finish();
-                saveNoStatus();
+                db.delTask(idno);
+                Toast.makeText(getApplicationContext(), "Your Task is Deleted" + idno, Toast.LENGTH_SHORT).show();
+                returnToMain();
             }
         });
 
 
-        Button save_button = findViewById(R.id.save_button);
+        Button update_button = findViewById(R.id.update_button);
 
-        save_button.setOnClickListener(new View.OnClickListener() {
+        update_button.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                String latitude = getIntent().getStringExtra("lat");
-                String longitude = getIntent().getStringExtra("longi");
-                String pname = getIntent().getStringExtra("pname");
-                String address = getIntent().getStringExtra("address");
-
-
-                EditText txtname = findViewById(R.id.title_edittext);
-                String task = txtname.getText().toString();
-
-                EditText todoTxt = findViewById(R.id.todo_edittext);
-                String todo = todoTxt.getText().toString();
-
-                SeekBar rad = findViewById(R.id.seekBar);
-                int radius = rad.getProgress();
-
-                EditText datetxt = findViewById(R.id.date);
-                String date = datetxt.getText().toString();
+                String utask = title.getText().toString();
+                String utodo = todo.getText().toString();
+                int uradius = radius.getProgress();
+                //      Integer.parseInt(date.getText().toString());
+                String udate = date.getText().toString();
 
                 // Validation
-                if (task.equals("")) {
+                if (utask.equals("")) {
                     Toast.makeText(getApplicationContext(), "Awww... Pleas give me a TITLE.", Toast.LENGTH_SHORT).show();
                     return;
                 }
-                if (date.equals("")) {
-                    Toast.makeText(getApplicationContext(), "Awww... do i have a DATE?.", Toast.LENGTH_SHORT).show();
+                if (udate.equals("")) {
+                    Toast.makeText(getApplicationContext(), "Awww... Do i have a DATE?.", Toast.LENGTH_SHORT).show();
                     return;
                 }
-
-                if (todo.equals("")) {
+                if (utodo.equals("")) {
                     Toast.makeText(getApplicationContext(), "Awww... I can not remind NOTHING.", Toast.LENGTH_SHORT).show();
                     return;
                 }
 
-                db.insertLocation(latitude, longitude, task, todo, radius, date, pname, address);
-                saveOkStatus();
 
-
+                db.updateTask(idno, utask, utodo, uradius, udate);
+                returnToMain();
             }
         });
-
-    }
-
-    @Override
-    public void onBackPressed() {
-        super.onBackPressed();
-        Intent intent = new Intent(this, MainActivity.class);
-        startActivity(intent);
-        this.finish();
     }
 
     private void onDatePickerDialog(){
@@ -121,7 +124,7 @@ public class TaskSetting extends AppCompatActivity {
         int mDay = c.get(Calendar.DAY_OF_MONTH); // current day
 
         // date picker dialog
-        datePickerDialog = new DatePickerDialog(TaskSetting.this,
+        datePickerDialog = new DatePickerDialog(ManageTask.this,
             new DatePickerDialog.OnDateSetListener() {
 
                 @Override
@@ -141,15 +144,15 @@ public class TaskSetting extends AppCompatActivity {
         datePickerDialog.show();
     }
 
-    private void saveOkStatus() {
-        Toast.makeText(getApplicationContext(), "Ahhh... Saved!! ", Toast.LENGTH_SHORT).show();
+    private void returnToMain() {
         Intent intent = new Intent(this, MainActivity.class);
         startActivity(intent);
         finish();
     }
 
-    private void saveNoStatus() {
-        Toast.makeText(getApplicationContext(), "Ahhh... NOT SAVED!! ", Toast.LENGTH_SHORT).show();
+    @Override
+    public void onBackPressed() {
+        super.onBackPressed();
         Intent intent = new Intent(this, MainActivity.class);
         startActivity(intent);
         this.finish();
