--- app/src/main/res/layout-v21/list_layout.xml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<GridLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-
-    <TextView
-        android:id="@+id/txttask"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_column="1"
-        android:layout_columnSpan="2"
-        android:layout_columnWeight="1"
-        android:layout_row="0"
-        android:text="task name"
-        android:textAppearance="@style/TextAppearance.AppCompat.Button" />
-
-
-    <TextView
-        android:id="@+id/txtdate"
-        android:layout_width="wrap_content"
-        android:layout_height="28dp"
-        android:layout_column="2"
-        android:layout_columnWeight="1"
-        android:layout_gravity="right"
-        android:layout_row="1"
-        android:text="date" />
-
-    <TextView
-        android:id="@+id/txtname"
-        android:layout_width="wrap_content"
-        android:layout_height="27dp"
-        android:layout_column="1"
-        android:layout_columnWeight="1"
-        android:layout_gravity="left|fill"
-        android:layout_row="1"
-        android:text="pname" />
-
-    <TextView
-        android:id="@+id/txtaddress"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_column="1"
-        android:layout_columnSpan="2"
-        android:layout_columnWeight="1"
-        android:layout_row="2"
-        android:text="Address" />
-
-</GridLayout>
