+++ app/src/main/res/layout/activity_about.xml
@@ -10,7 +10,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.ericabraham.leapfrog.MyMap">
+    tools:context="com.ericabraham.leapfrog.Ui.MyMap">
 
     <LinearLayout
         android:layout_width="match_parent"
