+++ app/src/main/res/layout/activity_main.xml
@@ -3,9 +3,9 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.ericabraham.leapfrog.MainActivity">
+    tools:context="com.ericabraham.leapfrog.Ui.MainActivity">
 
-    <include layout="@layout/content_main" />
+    <include layout="@layout/item_reminder_list" />
 
     <android.support.design.widget.FloatingActionButton
         android:id="@+id/fab"
@@ -13,6 +13,6 @@
         android:layout_height="wrap_content"
         android:layout_gravity="bottom|end"
         android:layout_margin="@dimen/fab_margin"
-        android:src="@drawable/ic_location_searching" />
+        android:src="@drawable/ic_location_searching"/>
 
 </android.support.design.widget.CoordinatorLayout>
