+++ app/src/main/res/layout/activity_manage_task.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.ericabraham.leapfrog.ManageTask">
+    tools:context="com.ericabraham.leapfrog.Ui.ManageTask">
 
 
     <ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
