+++ app/src/main/res/layout/activity_mymap.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.ericabraham.leapfrog.MyMap">
+    tools:context="com.ericabraham.leapfrog.Ui.MyMap">
 
     <!-- Line with Latitude and Longitude -->
     <LinearLayout
